SADistance

SADistance

Mapper2.SA.SADistance — Type.

abstract type SADistance

Fields

Documentation

Abstract distance type for placement.

API

getdistance

maxdistance

Basic Implementation

BasicDistance

Method List

source

API

Mapper2.SA.getdistance — Function.

getdistance(A::SADistance, a, b)

Return the distance from a to b with respect to A. Both a and b will implement getaddress.

Method List

getdistance(A, a, b)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Place/SA/Distance.jl:67.

source

Mapper2.SA.maxdistance — Function.

maxdistance(sa_struct, A::SADistance)

Return the maximum distance value of that occurs in sa_struct using the distance metric imposed by A.

Method List

maxdistance(sa_struct, A)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Place/SA/Distance.jl:69.

source

Implementations

Mapper2.SA.BasicDistance — Type.

struct BasicDistance{D} <: Mapper2.SA.SADistance

Fields

table

Simple look up table indexed by pairs of addresses. Returned value is the distance between the two addresses.

If an architecture has dimension "N", then the dimension of table is 2N.

Documentation

Basic implementation of SADistance. Constructs a look up table of distances between all address pairs. Addresses have a distance of 1 if there exists even one link between components at those addresses.

Method List

BasicDistance(table)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Place/SA/Distance.jl:56.

BasicDistance(toplevel, pathtable)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Place/SA/Distance.jl:76.

source