MapTables

MapTables

Mapper2.SA.AbstractMapTable — Type.

abstract type AbstractMapTable

Fields

Documentation

TODO

API

location_type

getlocations

isvalid

Implementations

MapTable

Method List

source

API

Mapper2.SA.location_type — Function.

Return the stored location type for a MapTable.

Method List

location_type(?)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Place/SA/MapTables.jl:96.

source

Mapper2.SA.getlocations — Function.

getlocations(maptable, class::Int)

Return a vector of locations that nodes of type class can occupy.

Method List

getlocations(maptable, class)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Place/SA/MapTables.jl:102.

source

Mapper2.SA.isvalid — Function.

isvalid(maptable, class, address :: Address)

Return true if nodes of type class can occupy address. In other words, there is some component at adddress that class can be mapped to.

Method List

isvalid(maptable, class, address)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Place/SA/MapTables.jl:111.

source

Implementations

Mapper2.SA.MapTable — Type.

struct MapTable{D} <: Mapper2.SA.AbstractMapTable

Fields

mask

Bit mask of whether a node class may be mapped to an address.

Accessing strategy: Look up the node class to index the outer vector. Index the inner array with an address.

A true entry means the task class can be mapped. A false entry means the task calss cannot be mapped.

Documentation

Default implementation of AbstractMapTable

Parameters:

D - The dimensionality of the Addresses in the table.

Method List

MapTable(mask)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Place/SA/MapTables.jl:67.

MapTable(toplevel, ruleset, equivalence_classes, pathtable)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Place/SA/MapTables.jl:89.

source