Core Types

Core Types

Port

Mapper2.MapperCore.Port — Type.

Port type for modeling input/output ports of a Component.

API

checkclass

invert

source

API

Mapper2.MapperCore.checkclass — Function.

checkclass(port, direction)

Return true if port is the correct class for the given Direction

source

Mapper2.MapperCore.invert — Function.

invert(port)

Return a version of port with the class inverted.

source

Mapper2.MapperCore.PortClass — Type.

Classification of port types.

source

Mapper2.MapperCore.Direction — Type.

Enum indicating a direction. Values: Source, Sink.

source

Link

Mapper2.MapperCore.Link — Type.

struct Link{P <: AbstractComponentPath}

Link data type for describing which ports are connected. Can have multiple sources and multiple sinks.

API

sources

dests

source

API

Mapper2.MapperCore.sources — Function.

sources(link)

Return Vector{Path{Port}} of sources for link.

source

Mapper2.MapperCore.dests — Function.

dests(link)

Return Vector{Path{Port}} of destinations for link.

source