Extensions

Extensions

The Mapper framework is designed to alter behavior mapping behavior using a certain set of core functions that are extended as needed for new subtypes of RuleSet. These methods only need to be extended if the default behavior is not sufficient.

In addition, placement and routing have types and methods that can be extended as well to more finely tune mapping.

Mapper2.MapperCore.isspecial — Function.

isspecial(ruleset::RuleSet, taskgraphnode::TaskgraphNode) :: Bool

Return true to disable move distance contraction for taskgraphnode during placement under ruleset.

Default: false

source

Mapper2.MapperCore.isequivalent — Function.

isequivalent(ruleset::RuleSet, a::TaskgraphNode, b::TaskgraphNode) :: Bool

Return true if TaskgraphNodes a and b are semantically equivalent for placement.

Default: true

source

Mapper2.MapperCore.ismappable — Function.

ismappable(ruleset::RuleSet, component::Component) :: Bool

Return true if some task can be mapped to component under ruleset.

Default: true

source

Mapper2.MapperCore.canmap — Function.

canmap(ruleset::RuleSet, t::TaskgraphNode, c::Component) :: Bool

Return true if t can be mapped to c under ruleset.

Default: true

source

Mapper2.MapperCore.canuse — Function.

canuse(ruleset::RuleSet, item::Union{Port,Link,Component}, edge::TaskgraphEdge)::Bool

Return true if edge can use item as a routing resource under ruleset.

Default: true

source

canuse(ruleset::RuleSet, link::RoutingLink, channel::RoutingChannel)::Bool

Return true if channel can be routed using link.

See: RoutingLink, RoutingChannel

Default: true

source

Mapper2.MapperCore.getcapacity — Function.

getcapacity(ruleset::RuleSet, item::Union{Port,Link,Component})

Return the capacity of routing resource item under ruleset.

Default: 1

source

Mapper2.MapperCore.is_source_port — Function.

is_source_port(ruleset::RuleSet, port::Port, edge::TaskgraphEdge)::Bool

Return true if port is a valid source port for edge under ruleset.

Default: true

source

Mapper2.MapperCore.is_sink_port — Function.

is_sink_port(ruleset::RuleSet, port::Port, edge::TaskgraphEdge)::Bool

Return true if port is a vlid sink port for edge under ruleset.

Default: true

source

Mapper2.MapperCore.needsrouting — Function.

needsrouting(ruleset::RuleSet, edge::TaskgraphEdge)::Bool

Return true if edge needs to be routed under ruleset.

Default: true

source

Mapper2.Routing.annotate — Function.

annotate(ruleset::RuleSet, item::Union{Port,Link,Component})

Return some <:RoutingLink for item. If item <: Component, it is a primitive. If not other primitives have been defined, it will be a mux.

See: BasicRoutingLink

Default: `BasicRoutingLink(capacity = getcapacity(ruleset, item))

source