MapperGraphs

MapperGraphs

Mapper2.MapperGraphs.SparseDiGraph — Type.

struct SparseDiGraph{T}

Fields

vertices

Documentation

Graph representation with arbitrary vertices of type T.

Method List

SparseDiGraph(vertices)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/MapperGraphs.jl:56.

source

Special Methods for SparseDiGraph

Mapper2.MapperGraphs.source_vertices — Function.

source_vertices(graph::SparseDiGraph{T}) :: Vector{T} where T

Return the vertices of graph that have no incoming edges.

source

Mapper2.MapperGraphs.sink_vertices — Function.

sink_vertices(graph::SparseDiGraph{T}) :: Vector{T} where T

Return the vertices of graph that have no outgoing edges.

source

Mapper2.MapperGraphs.make_lightgraph — Function.

make_lightgraph(graph::SparseDiGraph{T}) where T :: (SimpleDiGraph, Dict{T,Int})

Return tuple (g, d) where g :: SimpleDiGraph is a LightGraph ismorphic to graph and d maps vertices of graph to vertices of g.

source

Mapper2.MapperGraphs.linearize — Function.

linearize(graph::SparseDiGraph{T}) where T

Return a Vector{T} of vertices of graph in linearized traversal order if graph is linear. Throw error if ggraph is not a linear graph.

source