Routing Channels

Routing Channels

Mapper2.Routing.RoutingChannel — Type.

abstract type RoutingChannel

Fields

Documentation

Representation of channels in the taskgraph for routing.

API

start_vertices

stop_vertices

isless(a::RoutingChannel, b::RoutingChannel)

Implementations

BasicChannel

Method List

source

API

Base.isless — Method.

isless(a::RoutingChannel, b::RoutingChannel) :: Bool

Return true if a should be routed before b.

source

Implementations

Mapper2.Routing.BasicChannel — Type.

struct BasicChannel <: RoutingChannel

Fields

start_vertices

Direct storage for the Vector{PortVertices} of the sets of start vertices for each source of the channel.

stop_vertices

Direct storage for the Vector{PortVertices} of the sets of stop vertices for each source of the channel.

Documentation

Default implementation of RoutingChannel.

Method List

BasicChannel(start_vertices, stop_vertices)
BasicChannel(start_vertices, stop_vertices)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Route/Channels.jl:5.

source

Helpful Types

Mapper2.Routing.ChannelIndex — Type.

struct ChannelIndex

Fields

idx

Documentation

Type to access channels in the routing taskgraph. Essentially, it is just a wrapper for an integer, but typed to allow safer and clearer usage.

Method List

ChannelIndex(idx)
ChannelIndex(idx)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Route/Types.jl:8.

source

Mapper2.Routing.PortVertices — Type.

struct PortVertices

Fields

indices

Documentation

Indices of a RoutingGraph that can serve as either start or stop vertices (depending on the context) of one branch of a RoutingChannel.

Method List

PortVertices(indices)
PortVertices(indices)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Route/Types.jl:19.

source