Routing Graph

Routing Graph

Mapper2.Routing.RoutingGraph — Type.

struct RoutingGraph

Fields

graph

Adjacency information of routing resources, encode as a LightGraphs.SimpleDiGraph.

map

Translation information mapping elements on the parent TopLevel to indices in graph.

Implemented as a Dict{Path{<:Union{Port,Link,Component}}, Int64} where the values in the dict are the vertex index in graph of the key.

Documentation

Representation of the routing resources of a TopLevel.

Method List

RoutingGraph(graph, map)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Route/Graph.jl:26.

source

Mapper2.Routing.getmap — Function.

Return the map of a RoutingGraph

Method List

getmap(graph)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Route/Graph.jl:44.

source