Routing Links

Routing Links

The central type to represent routing resources during routing.

Mapper2.Routing.RoutingLink — Type.

abstract type RoutingLink

Fields

Documentation

Representation of routing resources in an architecture.

API

channels

cost

capacity

occupancy

addchannel

remchannel

Implementations

BasicRoutingLink - Reference this type for what methods of the API come for free when using various fields of the basic type.

Method List

source

API

Mapper2.Routing.channels — Function.

channels(link::RoutingLink) :: Vector{ChannelIndex}

Return list of channels currently occupying link.

source

Mapper2.Routing.cost — Function.

cost(link::RoutingLink) :: Float64

Return the base cost of a channel using link as a routing resource.

source

Mapper2.Routing.capacity — Function.

capacity(link::RoutingLink) :: Real

Return the capacity of link.

source

Mapper2.Routing.occupancy — Function.

occupancy(link::RoutingLink)

Return the number of channels currently using link.

source

Mapper2.Routing.addchannel — Function.

addchannel(link::RoutingLink, channel::ChannelIndex)

Record that channel is using link.

source

Mapper2.Routing.remchannel — Function.

remchannel(link::RoutingLink, channel::ChannelIndex)

Remove channel from the list of channels using link.

source

Implementations

Mapper2.Routing.BasicRoutingLink — Type.

struct BasicRoutingLink <: RoutingLink

Fields

channels

Vector of channels curently assigned to the link.

cost

Base cost of using this link.

capacity

Number of channels that can be mapped to this link without it being considered congested.

Documentation

Default implementation of RoutingLink

Simple container for channel indices, cost, and capacity.

Constructors

BasicRoutingLink(channels, cost, capacity)
BasicRoutingLink(channels, cost, capacity)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Route/Links.jl:12.

BasicRoutingLink(; cost, capacity)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Route/Links.jl:24.

Method List

BasicRoutingLink(channels, cost, capacity)
BasicRoutingLink(channels, cost, capacity)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Route/Links.jl:12.

BasicRoutingLink(; cost, capacity)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Route/Links.jl:24.

source