Routing

Routing

Mapper2.Routing.route! — Function.

route!(map::Map; meta_prefix = "")

Run pathfinder routing directly on map. Keyword argument meta_prefix allows controlling the prefix of the map.metadata keys generated below.

Records the following metrics into map.metadata:

routing_struct_time - Time it took to build the RoutingStruct

routing_struct_bytes - Memory allocation to build RoutingStruct

routing_passed :: Bool - true if routing passes check_routing, otherwise false.

routing_error :: Bool - true if routing experienced a connectivity error.

The following are also included if routing_error == false.

routing_time - Time to run routing.

routing_bytes - Memory allocation during routing.

routing_global_links - The number of global links used in the final routing.

source

Methods to Extend

Mapper2.Routing.routing_channel — Function.

routing_channel(ruleset::RuleSet, start, stop, edge::TaskgraphEdge)

Return <:RoutingChannel for edge. Arguments start and stop are return elements for start_vertices and stop_vertices respectively.

source