Taskgraph

Taskgraph

Mapper2.MapperCore.Taskgraph

Mapper2.MapperCore.TaskgraphEdge

Mapper2.MapperCore.TaskgraphNode

Mapper2.MapperCore.add_edge

Mapper2.MapperCore.add_node

Mapper2.MapperCore.getedge

Mapper2.MapperCore.getedges

Mapper2.MapperCore.getnode

Mapper2.MapperCore.getnodes

Mapper2.MapperCore.getsinks

Mapper2.MapperCore.getsources

Mapper2.MapperCore.hasnode

Mapper2.MapperCore.innode_names

Mapper2.MapperCore.innodes

Mapper2.MapperCore.nodenames

Mapper2.MapperCore.num_edges

Mapper2.MapperCore.num_nodes

Mapper2.MapperCore.outnode_names

Mapper2.MapperCore.outnodes

Mapper2.MapperCore.Taskgraph — Type.

struct Taskgraph

Fields

name

The name of the taskgraph

nodes

Nodes in the taskgraph. Type: Dict{String, TaskgraphNode}

edges

Edges in the taskgraph. Type: Vector{TaskgraphEdge}

node_edges_out

Outgoing adjacency list mapping node names to edge indices. Type: Dict{String, Vector{Int64}}

node_edges_in

Incoming adjacency list mapping node names to edge indices. Type: Dict{String, Vector{Int64}}

Documentation

Data structure encoding tasks and their relationships.

Method List

Taskgraph()
Taskgraph(name)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Taskgraphs.jl:70.

Taskgraph(name, nodes, edges)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Taskgraphs.jl:79.

Taskgraph(nodes, edges)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Taskgraphs.jl:129.

source

Important Types

Mapper2.MapperCore.TaskgraphNode — Type.

struct TaskgraphNode

Fields

name

The name of this task.

metadata

Documentation

Simple container representing a task in a taskgraph.

Method List

TaskgraphNode(name; metadata)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Taskgraphs.jl:10.

source

Mapper2.MapperCore.TaskgraphEdge — Type.

struct TaskgraphEdge

Fields

sources

Source task names.

sinks

Sink task names.

metadata

Documentation

Simple container representing an edge in a taskgraph.

Method List

TaskgraphEdge(source, sink; metadata)

defined at /home/travis/build/hildebrandmw/Mapper2.jl/src/Taskgraphs.jl:24.

source

API

Mapper2.MapperCore.add_edge — Function.

add_edge(taskgraph, edge)

Add a edge to taskgraph.

source

Mapper2.MapperCore.add_node — Function.

add_node(taskgraph, node)

Add a node to taskgraph. Error if node already exists.

source

Mapper2.MapperCore.getnodes — Function.

getnodes(taskgraph)

Return an iterator of TaskgraphNode yielding all nodes in taskgraph.

source

Mapper2.MapperCore.getnode — Function.

getnode(taskgraph, name)

Return the TaskgraphNode in taskgraph with name.

source

Mapper2.MapperCore.getedges — Function.

getedges(taskgraph)

Return an iterator of TaskgraphEdge yielding all edges in taskgraph.

source

Mapper2.MapperCore.getedge — Function.

getedge(taskgraph, index)

Return the TaskgraphEdge in taskgraph with index.

source

Mapper2.MapperCore.nodenames — Function.

nodenames(taskgraph)

Return an iterator yielding all names of nodes in taskgraph.

source

Mapper2.MapperCore.num_nodes — Function.

num_nodes(taskgraph)

Return the number of nodes in taskgraph.

source

Mapper2.MapperCore.num_edges — Function.

num_edges(taskgraph)

Return the number of edges in taskgraph.

source

Mapper2.MapperCore.getsources — Function.

getsources(taskgraph_edge)

Return the names of sources of a TaskgraphEdge.

source

getsources(taskgraph, edge)

Return Vector{TaskgraphNode} of sources for edge.

source

Mapper2.MapperCore.getsinks — Function.

getsinks(taskgraph_edge)

Return the names of sinks of a TaskgraphEdge.

source

getsinks(taskgraph, edge)

Return Vector{TaskgraphNode} of sinks for edge.

source

Mapper2.MapperCore.hasnode — Function.

hasnode(taskgraph, node)

Return true if taskgraph has a task named node.

source

Mapper2.MapperCore.innodes — Function.

innodes(taskgraph, node)

Return Vector{TaskgraphNode} of unique nodes that are the source of an edge ending at node.

source

Mapper2.MapperCore.innode_names — Function.

innode_names(taskgraph, node)

Return Set{String} of names of unique nodes that are the source of an edges ending at node.

source

Mapper2.MapperCore.outnodes — Function.

outnodes(taskgraph, node)

Return Vector{TaskgraphNode} of unique nodes that are the sink of an edge starting at node.

source

Mapper2.MapperCore.outnode_names — Function.

outnode_names(taskgraph, node)

Return Set{String} of names of unique nodes that are the sink of an edges starting at node.

source